feat(broker): 拉消息完成联调

+++ ccx-mq-broker/src/main/java/com/ccx/mq/broker/msg/StoreMsgInfo.java
@@ -10,7 +10,7 @@ import lombok.Data;
  */
 @Data
 public class StoreMsgInfo {
-    private int offset;
+    private Long offset;
     private String topic;
     private String msg;
 }
